progress

+++ kernel/src/drivers/mod.rs
@@ -11,6 +11,6 @@ pub fn init_drivers() {
     unsafe { fpu::init_fpu() };
     pci::init(); // pci init is blocking
     ahci::init(); // must be after pci
-    queue_spawn_kthread_named("keyboard", keyboard::driver_main);
-    queue_spawn_kthread_named("render", graphics::render_thread);
+    queue_spawn_kthread_named("keyboard", keyboard::driver_main as u64);
+    queue_spawn_kthread_named("render", graphics::render_thread as u64);
 }
